[fix] Equal size proportional sampling and memory leak
- Fixes #472

Here is the description that have been added to dataset loader on how
sampling works after this change:

Calculation of next batch is performed using following logic.

Current chosen iterator is selected based on the dataset probabilities set
in the change_dataloader function which is called everytime prepare_batch
is called.

If we get the next batch from iterator without any StopIteration exception,
we return it as it is. Otherwise, we have two cases:

1. In proportional sampling, since each dataset will have same number of epochs
at any given time, we need to yield StopIteration exception when all iterators
are finished. In turn, this will yield to __iter__ all reignite all
of the iterators. The code will not reach __iter__ until unless all iterators
are exhausted.

2. In the case of non-proportional sampling, epochs don't make sense.
Think of a case of equal proportion sampling for dataset x and y where x
is half the size of y. When x will complete its 2 epochs, y will have only
1 epoch completed. **So please don't use max_epochs or epoch based training
in this case as it won't be honored**. If an iterator is finished, we
just reignite it in this case and finished iterators variable isn't used.
This means that this case will never reach the __iter__ function ever
again.

Test Plan:

Tests have been added to check size proportional sampling and equal
sampling